- implement thread-safe path printing function sc_path_print() and use   it src/libopensc/ and src/pkcs15init/ - use size_t for the certlen parameter of sc_pkcs15init_update_certificate()
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@2811 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs15init/pkcs15-init.h
@@ -341,7 +341,7 @@ extern int	sc_pkcs15init_update_certificate(sc_pkcs15_card_t *p15card,
 				sc_profile_t *profile,
 				sc_pkcs15_object_t *obj,
 				const unsigned char *rawcert,
-				int certlen);
+				size_t certlen);
 
 extern int	sc_pkcs15init_create_file(struct sc_profile *,
 				struct sc_card *, struct sc_file *);
